feat: add try/catch to helper functions
